New Travis-CI Stage "Build"
This patch introduces a new stage to the Travis-CI configuration named
"build". It is responsible for building all the variations of the
REX-Ray binary, including a standard binary for each available driver.

Because the build phase uses containers instead of VMs, this new stage
will ensure any build-time that might occur due to the different
variations of build tags are caught before the overall build is put into
a potentially long wait queue due to requiring VM resources for
subsequent build stages.