API Throttling
This patch introduces API throttling. The feature depends on etcd for
persistent, external storage -- responsible for tracking active
connections and associating them with sustained and burst TTL leases.

The property "etcd.endpoints" or environment variable "ETCD_ENDPOINTS"
can be used to define one or more etcd endpoints. If the property is not
set then throttling is disabled.