Rewrite entire unit/init file on installation
When `rexray install` is called, it writes out a new unit/init script
depending on the system. This patch makes sure the file is erased
(truncated) first, so the entire file is written out a new. THe previous
logic merely overwrote the old file without erasing it first, which
could lead to incomplete/invalid files.