Do not check systemd status after stopping
A previous commit made it such that any error code returned from a
systemctl command led to an os.exit being called. This leads to errors
when doing package uninstalls in a variety of scenarios, because
the packages call `systemctl status -l rexray`, and if this is in an
error state for *any* reason (of which there can be several) the package
operation fails completely.

The fix here is temporary, and makes it such that a call to `status` is
not made after a service stop command is performed. The only visible to
effect to the user is less output if they happen to run
`rexray service stop|restart`, which is a command we will phase out
soon enough. `rexray service status` still behaves the same.