Don't use client provided NextDevice for AzureUD
The AzureUD driver was implemented using NextDevices, which allows a
client to specify what the next device (e.g. /dev/sdc) on the node is
where a block volume should be attached. The Controller code was not
using this in anyway, and always returning said device as the attach
token.

The Controller *was* querying the VM via the cloud provider API and
determing the next available device from that. These two sources are not
always consistent.

Instead, always use the next device info as it is determined from the
cloud provider's API, which is actually using LUNs.