Look at availability zones when searching for EFS mount targets
Before this commit the EFS driver would select a mount target based on the
subnet it is in. All mount targets in subnets not equal to the subnet of
the host would be ignored. This is incorrect behaviour as subnets do not
restrict access to a mount target. Now the mount target selection is
implemented by looking at the availability zone of the mount target
and host as these must be equal for the host to reach the mount target.